rename attribute_class module to attrs
